% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{BIC_cluster}
\alias{BIC_cluster}
\title{BIC function}
\usage{
BIC_cluster(data)
}
\arguments{
\item{data}{Data frame, last column is the cluster solution}
}
\value{
BIC of the given cluster solution
}
\description{
Calculates the BIC for a given cluster solution
}

