% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{funciongrafico}
\alias{funciongrafico}
\title{funciongrafico}
\usage{
funciongrafico(result, main = NULL)
}
\arguments{
\item{result}{Data frame, last column contains the cluster solution.}
}
\value{
A plot with the given cluster solution
}
\description{
Permite graficar el resultado de un algoritmo de cluster
(It needs to be improved)
}

