% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splitting.R
\name{functionSAR}
\alias{functionSAR}
\title{Function SAR}
\usage{
functionSAR(datos, minsize = ncol(datos) + log(nrow(datos) - ncol(datos)),
  r = ncol(datos) + log(nrow(datos) - ncol(datos)) - 1)
}
\arguments{
\item{datos}{Data frame, currently sagra supports only numeric variables.}

\item{minsize}{The minimum size of a cluster}

\item{r}{order of the considered discriminators}
}
\value{
A cluster solution
}
\description{
Splitting function from the SAR algorithm proposed by Peña and Tiao (2003)
This function calls to function SAR but also remove discriminators and group with sizes smaller than minsize parameter.
}

