% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{ordering}
\alias{ordering}
\title{ordering}
\usage{
ordering(data)
}
\arguments{
\item{data}{Data frame, the last column contains the cluster solution.}
}
\value{
A data frame with the cluster labels ordered.
}
\description{
Give ordered labels to the groups
The bigger group will be labelled as "1", and the rest are labelled by distance to "1" from closer to farest.
}

