% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sagra.R
\name{sagra}
\alias{sagra}
\title{Splitting And Group Recombining Algorithm (SAGRA)}
\usage{
sagra(data, minsize = ncol(data) + log(nrow(data) - ncol(data)))
}
\arguments{
\item{data}{Data frame, currently sagra supports only numeric variables.}

\item{minsize}{The minimum size of a cluster}
}
\value{
A cluster solution
}
\description{
This is a cluster algorithm proposed by Alvarez and Peña (2014)
Based on Bayes Factors
}

