% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombining.R
\name{testfunction}
\alias{testfunction}
\title{testfunction}
\usage{
testfunction(data)
}
\arguments{
\item{data}{Data frame, where the last column is the cluster solution}
}
\value{
Probability of $H_0$
}
\description{
Performs a Bayes Factor test for a given cluster solution
}

