% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{uni.plot}
\alias{uni.plot}
\title{uni.plot}
\usage{
uni.plot(x, symb = FALSE, quan = 1/2, alpha = 0.025, ...)
}
\arguments{
\item{x}{Matrix or data frame.}
}
\value{
A list with outliers and distances
}
\description{
This is a modification of the uni.plot function from the mvoutlier package
It calculates outliers without plotting
}

